+++ doc/WhatsNew.rst
@@ -6,6 +6,7 @@ Ver 4.1.0 (unreleased)
 ======================
 - change implementation of splash banner to a pop-up modal dialog
   with version string
+- fixed menubar integration on Mac OS X
 
 Ver 4.0.1 (2022-12-27)
 ======================